% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.formula.R
\name{reduce.formula}
\alias{reduce.formula}
\title{Reduce Formula}
\usage{
reduce.formula(dt, outcome.name, input.names, input.patterns = NA,
  max.input.categories = 20, max.outcome.categories.to.search = 4,
  return.as = "formula")
}
\arguments{
\item{dt}{TBD}

\item{outcome.name}{TBD}

\item{input.names}{TBD}

\item{input.patterns}{TBD}

\item{max.input.categories}{TBD}

\item{max.outcome.categories.to.search}{TBD}

\item{return.as}{TBD}
}
\description{
TBD
}
