% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.formula.R
\name{add.backtick}
\alias{add.backtick}
\title{Add backtick}
\usage{
add.backtick(x, include.backtick = TRUE)
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can be coerced by as.character to a character vector. Long vectors are supported.}

\item{include.backtick}{logical. If TRUE the corresponding component will remove any white space.}
}
\description{
Add backtick
}
