% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.formula.R
\name{add.backtick}
\alias{add.backtick}
\title{Add backticks if needed}
\usage{
add.backtick(x, include.backtick = "as.needed")
}
\arguments{
\item{x}{inputcharacter vector.}

\item{include.backtick}{Specifies whether a backtick should be added. Parameter values should be either 'all' or 'as.needed'}
}
\description{
Add backticks to make a appropriate variable.
}
