% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.backtick.R
\name{add.backtick}
\alias{add.backtick}
\title{Add backtick}
\usage{
add.backtick(x, include.backtick = "as.needed", dat = NULL)
}
\arguments{
\item{x}{Character value specifying the name of input parameters.}

\item{include.backtick}{specifies whether a backtick should be added. Parameter values should be either 'all' or 'as.needed'}

\item{dat}{Data}
}
\description{
Function that add backticks to the input variables.
}
