% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.formula.R
\name{add.backtick}
\alias{add.backtick}
\title{Add backticks if needed}
\usage{
add.backtick(x, include.backtick = "as.needed")
}
\arguments{
\item{x}{variables}

\item{include.backtick}{logical. If TRUE the corresponding component will remove any white space.}
}
\description{
Add backticks to make a appropriate variable.
}
