% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.formula.R
\name{reduce.existing.formula}
\alias{reduce.existing.formula}
\title{reduce existing formula}
\usage{
reduce.existing.formula(the.initial.formula, dat,
  max.input.categories = 20, max.outcome.categories.to.search = 4,
  order.as = "as.specified", include.backtick = "as.needed",
  format.as = "formula")
}
\arguments{
\item{the.initial.formula}{TBU}

\item{dat}{TBU}

\item{max.input.categories}{TBU}

\item{max.outcome.categories.to.search}{TBU}

\item{order.as}{TBU}

\item{include.backtick}{TBU}

\item{format.as}{TBU}
}
\description{
reduce existing formula
}
