% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.formula.R
\name{reduce.existing.formula}
\alias{reduce.existing.formula}
\title{reduce existing formula}
\usage{
reduce.existing.formula(dt, the.initial.formula,
  max.input.categories = 20, max.outcome.categories.to.search = 4,
  return.as = "formula")
}
\arguments{
\item{dt}{tBU}

\item{the.initial.formula}{TBU}

\item{max.input.categories}{TBU}

\item{max.outcome.categories.to.search}{TBU}

\item{return.as}{TBU}
}
\description{
reduce existing formula
}
